64bit pointer issue.
CVS patchset: 4561
CVS date: 2005/01/06 14:39:47

Mercurial revision: 0c3d75afdfc867f7c11a8e315dad3e0e5aa314ae

+++ lib/crm/cib/cib_client.c
@@ -486,7 +486,7 @@ gint ciblib_GCompareFunc(gconstpointer a, gconstpointer b)
 	if(a_client->callback == b_client->callback
 	   && safe_str_neq(a_client->event, b_client->event)) {
 		return 0;
-	} else if(((int)a_client->callback) < ((int)b_client->callback)) {
+	} else if(((long)a_client->callback) < ((long)b_client->callback)) {
 		return -1;
 	}
 	return 1;
